requirements.txt
accelerate==1.0.1
av==12.3.0
Brotli==1.0.9
certifi==2023.10.0
charset-normalizer==3.5.1
contourpy==1.1.1
cycler==0.12.1
fonttools==4.55.0
fsspec==2024.10.0
gmpy2==2.1.3
huggingface-hub==0.26.2
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.2
kiwisolver==1.4.7
MarkupSafe==2.1.4
matplotlib==3.7.5
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.26.1
opencv-python==4.10.0.84
packaging==24.2
peft==0.13.2
pillow==10.2.1
psutil==6.1.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
qwen-vl-utils==0.0.8
regex==2024.11.6
requests==2.31.0
safetensors==0.4.5
six==1.16.0
sympy==1.13
tokenizers==0.20.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.20.0
tqdm==4.67.0
transformers==4.46.2
triton==3.0.0
typing_extensions==4.8.0
urllib3==2.0.7
zipp==3.20.2